Refactored main app routing logic into dedicated subfolder.
